[Logical/Lazy] Tie `eval_if` and `eval` together, and update the docs
- Improve the documentation and tests of Lazy
- Add missing `operator!` to Logical's documentation
- In `eval_if`, evaluate branches with `eval` for more flexibility
- Use `lazy` instead of ugly workarounds in Sequence and Iterable

+++ include/boost/hana/lazy.hpp
@@ -47,7 +47,7 @@ namespace boost { namespace hana {
         { return detail::std::forward<Expr>(expr)(); }
 
         template <typename Expr>
-        static constexpr auto eval_helper(Expr&& expr, int)
+        static constexpr auto eval_helper(Expr&& expr, long)
             -> decltype(detail::std::forward<Expr>(expr)(hana::id))
         { return detail::std::forward<Expr>(expr)(hana::id); }
